requirements.txt
aiogram==2.24
httplib2==0.21.0
python-dotenv==0.21.1
emoji==2.2.0
gspread==5.7.2
pony==0.7.16
